requirements.txt
google-generativeai==0.7.2
pillow==10.4.0
flask==3.0.3